+++ news.md
@@ -1,6 +1,8 @@
 # 0.2.1.9000
 
 * Update authentication options for `scr_auth()` to include auto-authentication etc. from googleAuthR `0.6.2`
+* Added `searchAppearance` as a dimension option in `search_analytics()`
+
 
 # 0.2.1
 
@@ -11,7 +13,7 @@
 * Return an empty dataframe of NAs if no resutls in fetch instead of NULL
 * Include android-app check (#7)
 * Add `walk_data` parameter to `search_analytics` to get more data
-* Set default start and end dates in `search_analytics` to 93 days ago and 3 days ago respectivily.
+* Set default start and end dates in `search_analytics` to 93 days ago and 3 days ago respectively.
 * Correct bug for error in country code.  Will now return the 'Unknown Region' if not recognised (e.g. `CXX`)
 * Add `scr_auth` function that wraps `googleAuthR::gar_auth` so you don't need to load googleAuthR explicitly.
 
@@ -20,4 +22,4 @@
 * Move to using googleAuthR for authentication backend.
 
 ### 0.1.1
-* Change search_analytics() so if no dimensions will still return data, instead of NULL
+* Change `search_analytics()` so if no dimensions will still return data, instead of NULL